% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_user_add_blocks.R
\name{add_blocks}
\alias{add_blocks}
\title{Add blocks to a slimmr SLiM model.}
\usage{
add_blocks(slim_model, blocks_script = character(), after_block)
}
\arguments{
\item{slim_model}{An EidosModel object.}

\item{blocks_script}{A string of lines to add, e.g. from readLines().}

\item{after_block}{An integer index: which block to add block after?}
}
\description{
Adds one/more Eidos blocks (in order) to a slimmr SLiM model.
}
