% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_user_add_lines.R
\name{add_lines}
\alias{add_lines}
\title{Add lines to a slimmr SLiM model.}
\usage{
add_lines(slim_model, lines, after_line)
}
\arguments{
\item{slim_model}{An EidosModel object.}

\item{lines}{A character string of lines to add, e.g. from readLines().}

\item{after_line}{An integer index: which line to add line(s) after?}
}
\description{
Adds one/more Eidos line(s) to a slimmr SLiM model.
}
