% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_user_change_block_callback.R
\name{change_block_callback}
\alias{change_block_callback}
\title{Change the callback of a block in a slimmr SLiM model.}
\usage{
change_block_callback(slim_model, block_index, new_callback)
}
\arguments{
\item{slim_model}{An EidosModel object.}

\item{block_index}{Which block to change callback for?}

\item{new_callback}{E.g. "1000 late()"}
}
\description{
Changes the callback of one Eidos block in a slimmr SLiM model.
A more convenient alternative to slimmr::replace_eidos_pattern for callback
modification, but that function can be used to the same ends if you prefer.
}
