% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_user_duplicate_block.R
\name{duplicate_block}
\alias{duplicate_block}
\title{Duplicates a block in a slimmr SLiM model.}
\usage{
duplicate_block(slim_model, block_index, after_block)
}
\arguments{
\item{slim_model}{An EidosModel object.}

\item{block_index}{Integer index: block to duplicate.}

\item{after_block}{Integer index: which block to add duplicate block after?}
}
\description{
Duplicates one Eidos block in a slimmr SLiM model.
}
