% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_user.R
\name{inspect_eidos_script}
\alias{inspect_eidos_script}
\title{Inspect the Eidos script of a SLiM model.}
\usage{
inspect_eidos_script(slim_model)
}
\arguments{
\item{slim_model}{An EidosModel object.}
}
\description{
This functions outputs the script contained in an EidosModel
object to the console.
}
