% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_user_move_block.R
\name{move_block}
\alias{move_block}
\title{Move a block in a slimmr SLiM model to a different index.}
\usage{
move_block(slim_model, from_index, after_index)
}
\arguments{
\item{slim_model}{An EidosModel object.}

\item{from_index}{Original index of block.}

\item{after_index}{Move block to after block index ... (set 0 to move first!)}
}
\description{
Moves a block in a slimmr SLiM model to a different index.
}
