% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{newSlimModel}
\alias{newSlimModel}
\title{Function: newSlimModel}
\usage{
newSlimModel(
  description,
  scriptfile = NULL,
  generations = 1000,
  type = "WF",
  mu = 1e-08,
  rho = 1e-08,
  sex = NULL,
  dimensions = NULL,
  periodicity = NULL
)
}
\arguments{
\item{description}{A character description of the model.}

\item{scriptfile}{If using a prewritten model, the full path to the model. Otherwise NULL.}

\item{generations}{If creating model from scratch, integer value or constant symbol.}

\item{type}{If creating model from scratch, 'WF' or 'nonWF'.}

\item{mu}{If creating model from scratch, mutation rate.}

\item{rho}{If creating model from scratch, recombination rate.}

\item{sex}{If creating model from scratch, NULL, 'A', 'X', or 'Y'.}

\item{dimensions}{If creating model from scratch, NULL, 'XY', or 'XYZ'.}

\item{periodicity}{If creating model from scratch, 'NULL' or...}
}
\description{
Generate an object of class SlimModel.
}
