% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_user_remove_blocks.R
\name{remove_blocks}
\alias{remove_blocks}
\title{Remove blocks from a slimmr SLiM model.}
\usage{
remove_blocks(slim_model, block_indices)
}
\arguments{
\item{slim_model}{An EidosModel object.}

\item{block_indices}{Index/indices of block(s) to be removed.}
}
\description{
Removes one/more Eidos block(s) from a slimmr SLiM model.
}
