% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_user_remove_lines.R
\name{remove_lines}
\alias{remove_lines}
\title{Remove lines from a slimmr SLiM model.}
\usage{
remove_lines(slim_model, line_indices)
}
\arguments{
\item{slim_model}{An EidosModel object.}

\item{line_indices}{Index/indices of line(s) to be removed.}
}
\description{
Removes one/more Eidos line(s) from a slimmr SLiM model.
}
