% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_user_replace_eidos_pattern.R
\name{replace_eidos_pattern}
\alias{replace_eidos_pattern}
\title{Replace eidos patterns at model, block, or line level.}
\usage{
replace_eidos_pattern(
  slim_model,
  pattern,
  replacement,
  in_blocks = NULL,
  in_lines = NULL
)
}
\arguments{
\item{slim_model}{An EidosModel object.}

\item{pattern}{Eidos script pattern to replace.}

\item{replacement}{Eidos script replacement for the pattern.}

\item{in_blocks}{Which blocks to replace in? Incompatible with in_lines.
If in_blocks and in_lines are both NULL, will replace across entire model.}

\item{in_lines}{Which lines to replace in? Incompatible with in_blocks.
If in_blocks and in_lines are both NULL, will replace across entire model.}
}
\description{
Replaces an Eidos pattern at model, block, or line level.
}
