% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_user_run_slim.R
\name{run_slim}
\alias{run_slim}
\title{Run a SLiM model.}
\usage{
run_slim(
  model,
  reps = 1,
  seed = NULL,
  slim_output_verbosity = 1,
  output_parsing_function = function(x) return(x),
  parallel_nodes = 1,
  slim_command = "slim",
  output_runtime_diagnostics = FALSE,
  ...
)
}
\arguments{
\item{model}{An EidosModel object.}

\item{reps}{Number of replicate simulations.}

\item{seed}{Random seed for simulations}

\item{slim_output_verbosity}{An integer 0 - 2.}

\item{output_parsing_function}{A function applied to raw output from SLiM}

\item{parallel_nodes}{Number of load balancing parallel nodes to use.}

\item{slim_command}{Full path to SLiM executable. Default "slim" if on PATH.}

\item{output_runtime_diagnostics}{TRUE/FALSE. Output CPU and RAM usage?}

\item{...}{Named constant arguments to SLiM (i.e. passed to -d / -define)}
}
\value{
Either the raw output from SLiM, or that output as parsed by
output_parse_function.
}
\description{
This function runs a SLiM model.
}
