% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script.R
\name{Script}
\alias{Script}
\title{Class: Script}
\description{
Internal superclass for models and scriptblocks
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Script$new()}}
\item \href{#method-inspect}{\code{Script$inspect()}}
\item \href{#method-addlines}{\code{Script$addlines()}}
\item \href{#method-removelines}{\code{Script$removelines()}}
\item \href{#method-replacetext}{\code{Script$replacetext()}}
\item \href{#method-clone}{\code{Script$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Script$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inspect"></a>}}
\if{latex}{\out{\hypertarget{method-inspect}{}}}
\subsection{Method \code{inspect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Script$inspect(linenumshift = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addlines"></a>}}
\if{latex}{\out{\hypertarget{method-addlines}{}}}
\subsection{Method \code{addlines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Script$addlines(add, after, display = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removelines"></a>}}
\if{latex}{\out{\hypertarget{method-removelines}{}}}
\subsection{Method \code{removelines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Script$removelines(remove, display = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replacetext"></a>}}
\if{latex}{\out{\hypertarget{method-replacetext}{}}}
\subsection{Method \code{replacetext()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Script$replacetext(inlines = NULL, oldtext, newtext, display = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Script$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
