% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptBlock.R
\name{ScriptBlock}
\alias{ScriptBlock}
\title{Class: ScriptBlock}
\description{
Internal class to hold scriptblocks
}
\section{Super class}{
\code{\link[slimmr:Script]{slimmr::Script}} -> \code{ScriptBlock}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScriptBlock$new()}}
\item \href{#method-writein}{\code{ScriptBlock$writein()}}
\item \href{#method-writeout}{\code{ScriptBlock$writeout()}}
\item \href{#method-reindex}{\code{ScriptBlock$reindex()}}
\item \href{#method-clone}{\code{ScriptBlock$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="slimmr" data-topic="Script" data-id="addlines">}\href{../../slimmr/html/Script.html#method-addlines}{\code{slimmr::Script$addlines()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="slimmr" data-topic="Script" data-id="inspect">}\href{../../slimmr/html/Script.html#method-inspect}{\code{slimmr::Script$inspect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="slimmr" data-topic="Script" data-id="removelines">}\href{../../slimmr/html/Script.html#method-removelines}{\code{slimmr::Script$removelines()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="slimmr" data-topic="Script" data-id="replacetext">}\href{../../slimmr/html/Script.html#method-replacetext}{\code{slimmr::Script$replacetext()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScriptBlock$new(type, index, prewritten = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-writein"></a>}}
\if{latex}{\out{\hypertarget{method-writein}{}}}
\subsection{Method \code{writein()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScriptBlock$writein(newscript)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-writeout"></a>}}
\if{latex}{\out{\hypertarget{method-writeout}{}}}
\subsection{Method \code{writeout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScriptBlock$writeout()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reindex"></a>}}
\if{latex}{\out{\hypertarget{method-reindex}{}}}
\subsection{Method \code{reindex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScriptBlock$reindex(newindex)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScriptBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
