% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlimModel.R
\name{SlimModel}
\alias{SlimModel}
\title{Class: SlimModel}
\description{
Class: SlimModel

Class: SlimModel
}
\section{Super class}{
\code{\link[slimmr:Script]{slimmr::Script}} -> \code{SlimModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SlimModel$new()}}
\item \href{#method-print}{\code{SlimModel$print()}}
\item \href{#method-setup_genome}{\code{SlimModel$setup_genome()}}
\item \href{#method-setup_population}{\code{SlimModel$setup_population()}}
\item \href{#method-add_mutationtype}{\code{SlimModel$add_mutationtype()}}
\item \href{#method-add_interactiontype}{\code{SlimModel$add_interactiontype()}}
\item \href{#method-addblock}{\code{SlimModel$addblock()}}
\item \href{#method-inspectblock}{\code{SlimModel$inspectblock()}}
\item \href{#method-savetofile}{\code{SlimModel$savetofile()}}
\item \href{#method-run}{\code{SlimModel$run()}}
\item \href{#method-addlines}{\code{SlimModel$addlines()}}
\item \href{#method-removelines}{\code{SlimModel$removelines()}}
\item \href{#method-replacetext}{\code{SlimModel$replacetext()}}
\item \href{#method-moveblock}{\code{SlimModel$moveblock()}}
\item \href{#method-removeblock}{\code{SlimModel$removeblock()}}
\item \href{#method-clone}{\code{SlimModel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="slimmr" data-topic="Script" data-id="inspect">}\href{../../slimmr/html/Script.html#method-inspect}{\code{slimmr::Script$inspect()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$new(
  description,
  scriptfile = NULL,
  generations = 1000,
  type = "WF",
  mu = 1e-08,
  rho = 1e-08,
  sex = NULL,
  dimensions = NULL,
  periodicity = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup_genome"></a>}}
\if{latex}{\out{\hypertarget{method-setup_genome}{}}}
\subsection{Method \code{setup_genome()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$setup_genome()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup_population"></a>}}
\if{latex}{\out{\hypertarget{method-setup_population}{}}}
\subsection{Method \code{setup_population()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$setup_population()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_mutationtype"></a>}}
\if{latex}{\out{\hypertarget{method-add_mutationtype}{}}}
\subsection{Method \code{add_mutationtype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$add_mutationtype()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_interactiontype"></a>}}
\if{latex}{\out{\hypertarget{method-add_interactiontype}{}}}
\subsection{Method \code{add_interactiontype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$add_interactiontype()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addblock"></a>}}
\if{latex}{\out{\hypertarget{method-addblock}{}}}
\subsection{Method \code{addblock()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$addblock(index, blocktype, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inspectblock"></a>}}
\if{latex}{\out{\hypertarget{method-inspectblock}{}}}
\subsection{Method \code{inspectblock()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$inspectblock(index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-savetofile"></a>}}
\if{latex}{\out{\hypertarget{method-savetofile}{}}}
\subsection{Method \code{savetofile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$savetofile(path, filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$run(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addlines"></a>}}
\if{latex}{\out{\hypertarget{method-addlines}{}}}
\subsection{Method \code{addlines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$addlines(add, after)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removelines"></a>}}
\if{latex}{\out{\hypertarget{method-removelines}{}}}
\subsection{Method \code{removelines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$removelines(remove)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replacetext"></a>}}
\if{latex}{\out{\hypertarget{method-replacetext}{}}}
\subsection{Method \code{replacetext()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$replacetext(inlines = NULL, oldtext, newtext)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-moveblock"></a>}}
\if{latex}{\out{\hypertarget{method-moveblock}{}}}
\subsection{Method \code{moveblock()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$moveblock(index, to)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removeblock"></a>}}
\if{latex}{\out{\hypertarget{method-removeblock}{}}}
\subsection{Method \code{removeblock()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$removeblock(index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlimModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
