% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimmr_v1.0_user.R
\name{import_slim_model}
\alias{import_slim_model}
\title{Import a SLiM model from an Eidos script.}
\usage{
import_slim_model(script_path, name = "")
}
\arguments{
\item{script_path}{Path to .slim script.}

\item{name}{A descriptive name for the model.}
}
\value{
An EidosModel object.
}
\description{
This functions import a .slim script file and converts it to a
slimmr EidosModel object. This object can be manipulated and run with the
various functions of slimmr.
}
